refactor: Move save unpacker/repacker classes to save utils class
